visual improvements in the script metadata tab

+++ AWS/app/aws/configure/script/ScriptManagerCtrl.js
@@ -1,5 +1,5 @@
 var scriptUploaded;
-var scriptModule = angular.module('aws.configure.script', ['ngGrid', 'mk.editablespan']).controller("ScriptManagerCtrl", function($scope, scriptManagerService, queryService) {
+var scriptModule = angular.module('aws.configure.script', ['ngGrid', 'mk.editablespan']).controller("ScriptManagerCtrl", function($scope, $modal, scriptManagerService, queryService) {
 
 	  $scope.service = scriptManagerService;
 	  $scope.queryService = queryService;
@@ -12,37 +12,44 @@ var scriptModule = angular.module('aws.configure.script', ['ngGrid', 'mk.editabl
 	  $scope.status = "";
 	  $scope.statusColor = "#3276B1";
 	  $scope.jsonbtn = "json";
-	  
+	  $scope.EditDone = "Edit";
 	  $scope.inputsAsString = "";
 	  
 	  $scope.rScriptListOptions = {
 			  data: 'rScripts',
-			  columnDefs: [{field: 'Script', displayName: 'R Scripts'}],
+			  columnDefs: [{field: 'Script', displayName: ''}],
 			  selectedItems: $scope.selectedScript,
 			  multiSelect: false,
-			  enableRowSelection: true
+			  enableRowSelection: true,
+			  headerRowHeight:0,
+			  keepLastSelected : false,
 	      };
 	      
 	  $scope.stataScriptListOptions = {
 			  data: 'stataScripts',
-			  columnDefs: [{field: 'Script', displayName: 'Stata Scripts'}],
+			  columnDefs: [{field: 'Script', displayName: ''}],
 			  selectedItems: $scope.selectedScript,
 			  multiSelect: false,
-			  enableRowSelection: true
+			  enableRowSelection: true,
+			  headerRowHeight:0,
+			  keepLastSelected : false,
 	  };
 	  
 	  $scope.scriptMetadataGridOptions = {
 			  data: 'scriptMetadata.inputs',
-		  columnDefs : [{field : "param", displayName : "Parameter"},
-		               {field :"type", displayName : "Type", cellTemplate : '<select  ng-input="COL_FIELD" ng-model="COL_FIELD" ng-options="input for input in inputTypes" style="align:center"></select>'},
-		               {field : "columnType", displayName : "Column Type", cellTemplate : '<select  ng-input="COL_FIELD" ng-if="scriptMetadata.inputs[row.rowIndex].type == &quot;column&quot;" ng-model="COL_FIELD" ng-options="type for type in columnTypes" style="align:center"></select>'},
+			  columnDefs : [{field : "param", displayName : "Parameter"},
+		               {field :"type", displayName : "Type", enableCellEdit : false, cellTemplate : '<select  ng-input="COL_FIELD" ng-model="COL_FIELD" ng-options="input for input in inputTypes" style="align:center"></select>'},
+		               {field : "columnType", displayName : "Column Type", enableCellEdit : false, cellTemplate : '<select  ng-input="COL_FIELD" ng-if="scriptMetadata.inputs[row.rowIndex].type == &quot;column&quot;" ng-model="COL_FIELD" ng-options="type for type in columnTypes" style="align:center"></select>'},
 		               {field : "options", displayName : "Options"},
 		               {field : "default", displayName : "Default"},
 		               {field : "description", displayName : "Description"}],
 			  multiSelect: false,
 			  enableRowSelection: true,
+			  keepLastSelected : false,
+			  enableCellEditOnFocus: true,
 			  enableCellEdit : true,
-			  selectedItems : $scope.selectedRow
+			  selectedItems : $scope.selectedRow,
+			  enableSorting : false,
 	  };
 	  
 	  $scope.inputTypes = ["column", "options", "boolean", "value", "multiColumns", ""];
@@ -91,6 +98,7 @@ var scriptModule = angular.module('aws.configure.script', ['ngGrid', 'mk.editabl
 		
 		// every time editScript is turnOff, we should save the changes.
 		if(!$scope.editScript) {
+			$scope.EditDone = "Edit";
 			if($scope.script.content && $scope.selectedScript[0].Script) {
 				scriptManagerService.saveScriptContent($scope.selectedScript[0].Script,  $scope.script.content).then(function (result) {
 					if(result) {
@@ -98,6 +106,8 @@ var scriptModule = angular.module('aws.configure.script', ['ngGrid', 'mk.editabl
 					}
 				});
 			}
+		} else {
+			$scope.EditDone = "Done";
 		}
 	  };
 	  
@@ -119,7 +129,7 @@ var scriptModule = angular.module('aws.configure.script', ['ngGrid', 'mk.editabl
 	  });
 	  $scope.$watch('scriptMetadata.inputs', function () {
 		  if($scope.scriptMetadata.inputs) {
-			  $scope.inputsAsString = angular.toJson($scope.scriptMetadata.inputs); 
+			  $scope.inputsAsString = angular.toJson($scope.scriptMetadata.inputs, true); 
 		  }
 	  }, true);
 	  /***********************/
@@ -129,30 +139,40 @@ var scriptModule = angular.module('aws.configure.script', ['ngGrid', 'mk.editabl
 	 };
 	
 	 $scope.removeRow = function() {
+		 if($scope.viewasjson)
+			 return;
 		 if($scope.scriptMetadataGridOptions.selectedItems.length) {
 			 var index = $scope.scriptMetadata.inputs.indexOf($scope.scriptMetadataGridOptions.selectedItems[0]);
 			 $scope.scriptMetadata.inputs.splice(index, 1);
+			 $scope.scriptMetadataGridOptions.selectAll(false);
 		 }
 	 };
 	 
 	 $scope.deteleScript = function () {
 		 if($scope.selectedScript.length) {
-			scriptManagerService.deleteScript($scope.selectedScript[0].Script).then(function(status) {
-				if(status) {
-					console.log("script deleted successfully");
-			}
-			$scope.selectedScript[0].Script = "";
-			$scope.script.content = "";
-			$scope.scriptMetadata = {};
-			refreshScripts();
-			});
+			 if($scope.selectedScript[0].Script) {
+				 scriptManagerService.deleteScript($scope.selectedScript[0].Script).then(function(status) {
+					 if(status) {
+						 console.log("script deleted successfully");
+						 $scope.rScriptListOptions.selectAll(false);
+						 $scope.stataScriptListOptions.selectAll(false);
+					 }
+					 $scope.selectedScript.Script = "";
+					 $scope.script.content = "";
+					 $scope.scriptMetadata = {};
+					 refreshScripts();
+				 });
+			 }
 		 }
 	 };
+	 $scope.$watch('selectedScript.length', function () {
+		 console.log($scope.selectedScript.length);
+	 });
 	 
 	 $scope.saveChanges = function () {
 		 if($scope.selectedScript[0])
 		 {
-			 scriptManagerService.saveScriptMetadata($scope.selectedScript[0].Script, angular.toJson($scope.scriptMetadata)).then(function(result) { 
+			 scriptManagerService.saveScriptMetadata($scope.selectedScript[0].Script, angular.toJson($scope.scriptMetadata, true)).then(function(result) { 
 				 
 				 if(result) {
 					 refreshScripts();
@@ -163,43 +183,42 @@ var scriptModule = angular.module('aws.configure.script', ['ngGrid', 'mk.editabl
 						 $scope.$apply();
 					 }, 1000);
 				 } else {
-					 $scope.statusColor = "green";
-					 $scope.status = "Changes successful!";
+					 $scope.statusColor = "red";
+					 $scope.status = "Changes unsuccessful!";
 				 }
 			 });
 		 }
 	 };
-});
-scriptModule.controller('AddScriptDialogController', function ($scope, $dialog, scriptManagerService, queryService) {
-	$scope.opts = {
-		 backdrop: false,
-         backdropClick: true,
-         dialogFade: true,
-         keyboard: true,
-         templateUrl: 'aws/configure/script/uploadNewScript.html',
-         controller: 'AddScriptDialogInstanceCtrl',
-	};
-	
     $scope.saveNewScript = function (content, metadata) {
-      
-    	var d = $dialog.dialog($scope.opts);
-    	d.open();
+    	$modal.open({
+			 backdrop: false,
+	         backdropClick: true,
+	         dialogFade: true,
+	         keyboard: true,
+	         templateUrl: 'aws/configure/script/uploadNewScript.html',
+	         controller: 'AddScriptDialogInstanceCtrl',
+		});
     };
-    
- })
- .controller('AddScriptDialogInstanceCtrl', function ($rootScope, $scope, dialog, scriptManagerService) {
+}).controller('AddScriptDialogInstanceCtrl', function ($rootScope, $scope, $modalInstance, scriptManagerService) {
 	  
 	 $scope.fileName = "";
 	 $scope.metadata = "";
+	 $scope.validText = "";
+	 $scope.isScriptValid = false;
 	 $scope.step = 1;
 	 
-	 $scope.scriptUploaded = {};
+	 $scope.scriptUploaded = {
+			 script : {
+				 filename : "",
+				 content :  ""
+			 }
+	 };
 	 $scope.metadataUploaded = {
 			 content : ""
 	 };
 	
 	 $scope.close = function () {
-	  dialog.close();
+		 $modalInstance.close();
 	 };
 	 
 	 $scope.uploadSuccessful = "na";
@@ -224,6 +243,17 @@ scriptModule.controller('AddScriptDialogController', function ($scope, $dialog, 
 			  enableCellEdit : false,
 	  };
 	 
+	$scope.$watch('scriptUploaded.script.filename', function () {
+		scriptManagerService.scriptExists($scope.scriptUploaded.script.filename).then(function(result) {
+			if(result) {
+				$scope.validText = "script already exists";
+				$scope.isScriptValid = false;
+			} else {
+				$scope.validText = "";
+				$scope.isScriptValid = true;
+			}
+		});
+	});
 	$scope.$watch('step', function(n, o) { 
 		if(n == 3) {
 			var scriptName = "";
